+++ src/endpoints/map/stops.cc
@@ -3,11 +3,11 @@
 #include "net/bad_request_exception.h"
 #include "net/too_many_exception.h"
 
-#include "motis/display_filter.h"
 #include "motis/parse_location.h"
 #include "motis/place.h"
 #include "motis/timetable/modes_to_clasz_mask.h"
 
+namespace n = nigiri;
 namespace motis::ep {
 
 api::stops_response stops::operator()(boost::urls::url_view const& url) const {
@@ -16,6 +16,10 @@ api::stops_response stops::operator()(boost::urls::url_view const& url) const {
   auto const max = parse_location(query.max_);
   auto const modes = query.modes_;
   auto const grouped = query.grouped_.value_or(false);
+
+  auto any_allowed = [](auto const& mask, auto const& cs) -> bool {
+    return (mask & cs) != 0;
+  };
 
   utl::verify<net::bad_request_exception>(
       min.has_value(), "min not a coordinate: {}", query.min_);
@@ -34,7 +38,7 @@ api::stops_response stops::operator()(boost::urls::url_view const& url) const {
     auto const mask = modes.has_value() ? to_clasz_mask(*modes)
                                         : n::routing::all_clasz_allowed();
 
-    if (n::routing::any_allowed(mask, clasz)) {
+    if (any_allowed(mask, clasz)) {
       auto const [_, inserted] = seen_places.insert(place_idx);
       if (!grouped || inserted) {
         res.emplace_back(to_place(&tt_, &tags_, w_, pl_, matches_, ae_, tz_,
